refactor(State): rename Status -> State
