+++ src/main/java/blackjack/domain/state/Bust.java
@@ -1,9 +1,9 @@
-package blackjack.domain.status;
+package blackjack.domain.state;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.participant.Cards;
 
-public class Bust implements Status{
+public class Bust implements State {
 
     private final Cards cards;
 
@@ -12,7 +12,7 @@ public class Bust implements Status{
     }
 
     @Override
-    public Status draw(Card card) {
+    public State draw(Card card) {
         throw new IllegalStateException("Bust 상태에선 카드를 받을 수 없습니다.");
     }
