+++ src/test/java/blackjack/domain/state/ReadyTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.status;
+package blackjack.domain.state;
 
 import static blackjack.domain.card.CardNumber.*;
 import static blackjack.domain.card.CardSymbol.*;
@@ -18,53 +18,53 @@ public class ReadyTest {
     @DisplayName("draw으로 카드 한 장을 뽑은 경우 Ready 상태이다.")
     void draw() {
         //given
-        Status status = new Ready();
+        State state = new Ready();
 
         //when
-        Status newStatus = status.draw(new Card(HEART, JACK));
+        State newState = state.draw(new Card(HEART, JACK));
 
         //then
-        assertThat(newStatus).isInstanceOf(Ready.class);
+        assertThat(newState).isInstanceOf(Ready.class);
     }
 
     @Test
     @DisplayName("2 장을 뽑으면 Hit 상태로 바뀐다.")
     void toHit() {
         //given
-        Status status = new Ready();
+        State state = new Ready();
 
         //when
-        Status newStatus = status.draw(new Card(HEART, JACK));
-        Status hitStatus = newStatus.draw(new Card(CLUB, JACK));
+        State newState = state.draw(new Card(HEART, JACK));
+        State hitState = newState.draw(new Card(CLUB, JACK));
 
         //then
-        assertThat(hitStatus).isInstanceOf(Hit.class);
+        assertThat(hitState).isInstanceOf(Hit.class);
     }
 
     @Test
     @DisplayName("처음 2장의 카드의 합이 21이면 Blackjack 상태로 바뀐다.")
     void toBlackjack() {
         //given
-        Status status = new Ready();
+        State state = new Ready();
 
         //when
-        Status newStatus = status.draw(new Card(HEART, JACK));
-        Status blackjackStatus = newStatus.draw(new Card(HEART, ACE));
+        State newState = state.draw(new Card(HEART, JACK));
+        State blackjackState = newState.draw(new Card(HEART, ACE));
 
         //then
-        assertThat(blackjackStatus).isInstanceOf(Blackjack.class);
+        assertThat(blackjackState).isInstanceOf(Blackjack.class);
     }
 
     @Test
     @DisplayName("Hit 상태에서 draw 로 21이 초과하면 Bust 상태로 바뀐다.")
     void toBust() {
         //given
-        Status hitStatus = new Hit(new Cards(Set.of(new Card(HEART, JACK), new Card(CLUB, JACK))));
+        State hitState = new Hit(new Cards(Set.of(new Card(HEART, JACK), new Card(CLUB, JACK))));
 
         //when
-        Status bustStatus = hitStatus.draw(new Card(HEART, TWO));
+        State bustState = hitState.draw(new Card(HEART, TWO));
 
         //then
-        assertThat(bustStatus).isInstanceOf(Bust.class);
+        assertThat(bustState).isInstanceOf(Bust.class);
     }
 }
